refactor experiment page
