+++ submarine-workbench/workbench-web/src/app/pages/workbench/experiment/experiment-home/experiment-home.component.html
@@ -1,21 +1,51 @@
-<!--
- Licensed to the Apache Software Foundation (ASF) under one
- or more contributor license agreements.  See the NOTICE file
- distributed with this work for additional information
- regarding copyright ownership.  The ASF licenses this file
- to you under the Apache License, Version 2.0 (the
- "License"); you may not use this file except in compliance
- with the License.  You may obtain a copy of the License at
+<div style="margin: 15px; padding: 15px; background-color: white">
+  <div align="right">
+    <nz-radio-group>
+      <label nz-radio-button nzValue="All">All</label>
+      <label nz-radio-button nzValue="Own">Owned By Me</label>
+      <label nz-radio-button nzValue="Access">Accessible By Me</label>
+    </nz-radio-group>
+    <nz-input-group
+      nzSearch
+      style="width: 300px; margin-top: 15px; margin-left: 10px; margin-right: 5px"
+      [nzAddOnAfter]="suffixIconButton"
+    >
+      <input type="text" nz-input placeholder="input search text" />
+    </nz-input-group>
+    <ng-template #suffixIconButton>
+      <button nz-button nzType="primary" nzSearch><i nz-icon nzType="search"></i></button>
+    </ng-template>
 
-   http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing,
- software distributed under the License is distributed on an
- "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- KIND, either express or implied.  See the License for the
- specific language governing permissions and limitations
- under the License.
--->
-
-<!-- <p>experiment-home works!</p> -->
-<!-- TODO: should split home page from experiment root components -->
+    <button
+      nz-button
+      id="openExperiment"
+      nzType="primary"
+      style="margin-right: 5px; margin-bottom: 15px; margin-top: 15px"
+      (click)="form.initModal('create')"
+    >
+      <i nz-icon nzType="plus"></i>
+      New Experiment
+    </button>
+    <button
+      nz-button
+      nzType="primary"
+      style="margin-bottom: 15px; margin-top: 15px"
+      nz-popconfirm
+      nzTitle="Confirm to delete?"
+      nzCancelText="Cancel"
+      nzOkText="Ok"
+      (nzOnConfirm)="deleteExperiments()"
+    >
+      Delete
+    </button>
+  </div>
+  <submarine-experiment-list
+    [experimentList]="experimentList"
+    [checkedList]="checkedList"
+    [isLoading]="isLoading"
+    (deleteExperiment)="onDeleteExperiment($event, true)"
+    (initModal)="onInitModal($event)"
+    [(selectAllChecked)]="selectAllChecked"
+  ></submarine-experiment-list>
+  <submarine-experiment-form #form></submarine-experiment-form>
+</div>
