+++ submarine-workbench/workbench-web/src/app/pages/workbench/experiment/experiment-home/experiment-list/experiment-list.component.scss
@@ -1,4 +1,4 @@
-/*!
+/**
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
  * distributed with this work for additional information
@@ -7,7 +7,7 @@
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
  *
- *   http://www.apache.org/licenses/LICENSE-2.0
+ * http: //www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
@@ -16,21 +16,3 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
-#form-type-container {
-  display: grid;
-  grid-auto-rows: 5rem;
-  grid-template-columns: repeat(auto-fit, minmax(10rem, 1fr));
-  column-gap: 3rem;
-  & button {
-    height: 100%;
-  }
-}
-
-input.ng-invalid.ng-touched {
-  border: 1px solid red;
-}
-
-textarea.ng-invalid.ng-touched {
-  border: 1px solid red;
-}
