+++ submarine-workbench/workbench-web/src/app/pages/workbench/experiment/experiment.module.ts
@@ -1,20 +1,3 @@
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { RouterModule } from '@angular/router';
-import { PipeSharedModule } from '@submarine/pipe/pipe-shared.module';
-import { NgxChartsModule } from '@swimlane/ngx-charts';
-import { NgZorroAntdModule } from 'ng-zorro-antd';
-import { ExperimentCustomizedFormComponent } from './experiment-customized-form/experiment-customized-form.component';
-import { ChartsComponent } from './experiment-info/charts/charts.component';
-import { ExperimentInfoComponent } from './experiment-info/experiment-info.component';
-import { HyperParamsComponent } from './experiment-info/hyper-params/hyper-params.component';
-import { MetricsComponent } from './experiment-info/metrics/metrics.component';
-import { OutputsComponent } from './experiment-info/outputs/outputs.component';
-import { ExperimentComponent } from './experiment.component';
-import { ExperimentRoutingModule } from './experiment-routing.module';
-import { ExperimentHomeComponent } from './experiment-home/experiment-home.component';
-import { ExperimentPredefinedFormComponent } from './experiment-predefined-form/experiment-predefined-form.component';
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
@@ -34,6 +17,27 @@ import { ExperimentPredefinedFormComponent } from './experiment-predefined-form/
  * under the License.
  */
 
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { RouterModule } from '@angular/router';
+import { PipeSharedModule } from '@submarine/pipe/pipe-shared.module';
+import { NgxChartsModule } from '@swimlane/ngx-charts';
+import { NgZorroAntdModule } from 'ng-zorro-antd';
+import { ExperimentComponent } from './experiment.component';
+import { ExperimentListComponent } from './experiment-home/experiment-list/experiment-list.component';
+import { ExperimentRoutingModule } from './experiment-routing.module';
+import { ExperimentInfoComponent } from './experiment-info/experiment-info.component';
+import { HyperParamsComponent } from './experiment-info/hyper-params/hyper-params.component';
+import { MetricsComponent } from './experiment-info/metrics/metrics.component';
+import { ChartsComponent } from './experiment-info/charts/charts.component';
+import { OutputsComponent } from './experiment-info/outputs/outputs.component';
+import { ExperimentHomeComponent } from './experiment-home/experiment-home.component';
+import { ExperimentService } from '@submarine/services/experiment.service';
+import { ExperimentFormComponent } from './experiment-home/experiment-form/experiment-form.component';
+import { ExperimentPredefinedFormComponent } from './experiment-home/experiment-form/experiment-predefined-form/experiment-predefined-form.component';
+import { ExperimentCustomizedFormComponent } from './experiment-home/experiment-form/experiment-customized-form/experiment-customized-form.component';
+
 @NgModule({
   exports: [ExperimentComponent],
   imports: [
@@ -44,18 +48,21 @@ import { ExperimentPredefinedFormComponent } from './experiment-predefined-form/
     NgxChartsModule,
     RouterModule,
     PipeSharedModule,
-    ExperimentRoutingModule
+    ExperimentRoutingModule,
   ],
+  providers: [ExperimentService],
   declarations: [
     ExperimentComponent,
+    ExperimentListComponent,
     ExperimentInfoComponent,
     HyperParamsComponent,
     MetricsComponent,
     ChartsComponent,
     OutputsComponent,
-    ExperimentCustomizedFormComponent,
     ExperimentHomeComponent,
-    ExperimentPredefinedFormComponent
-  ]
+    ExperimentFormComponent,
+    ExperimentPredefinedFormComponent,
+    ExperimentCustomizedFormComponent,
+  ],
 })
 export class ExperimentModule {}
